impleted websockets through pusher

+++ src/app/api/strikes/route.ts
@@ -1,3 +1,4 @@
+import { pusherServer } from '@/lib/pusher/server';
 import { getCurrentUser } from '@/lib/auth/currentUser';
 import { addStrike, deleteStrike, getStrikeHistory, updateStrike } from '@/lib/strikes';
 
@@ -5,6 +6,11 @@ export async function POST(request: Request) {
   const { pledgeId, createdById, amount, reason } = await request.json();
 
   const newStrikeEvent = await addStrike(pledgeId, createdById, amount, reason);
+
+  await pusherServer.trigger('private-strikes', 'strike-created', {
+    pledgeId,
+    userId: createdById
+  });
 
   return Response.json(newStrikeEvent);
 }
@@ -34,18 +40,32 @@ export async function DELETE(request: Request) {
   const { searchParams } = new URL(request.url);
 
   const strikeId = searchParams.get('strikeId');
+  const pledgeId = searchParams.get('pledgeId');
+  const deletedById = searchParams.get('deletedById');
 
   if (!strikeId) return Response.json({ error: 'No strike id.' });
+  if (!pledgeId) return Response.json({ error: 'No pledge id.' });
+  if (!deletedById) return Response.json({ error: 'No user id.' });
 
   const deletedStrike = await deleteStrike(strikeId);
+
+  await pusherServer.trigger('private-strikes', 'strike-deleted', {
+    pledgeId,
+    userId: deletedById
+  });
 
   return Response.json(deletedStrike);
 }
 
 export async function PUT(request: Request) {  
-  const { id, amount, reason } = await request.json();
+  const { id, amount, reason, pledgeId, updatedById } = await request.json();
 
   await updateStrike(id, amount, reason);
+
+  await pusherServer.trigger('private-strikes', 'strike-updated', {
+    pledgeId,
+    userId: updatedById
+  });
 
   return Response.json({ success: true });
 }
